Update DTO mappers for request
