+++ src/main/java/de/samply/bbmri/negotiator/db/util/DbUtil.java
@@ -48,18 +48,13 @@ import eu.bbmri.eric.csit.service.negotiator.sync.directory.dto.DirectoryNetwork
 import org.jooq.*;
 import org.jooq.Record;
 import org.jooq.exception.DataAccessException;
-import org.jooq.impl.DSL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 
 import de.samply.bbmri.negotiator.Config;
 import de.samply.bbmri.negotiator.NegotiatorConfig;
 import de.samply.bbmri.negotiator.jooq.Tables;
 import de.samply.bbmri.negotiator.jooq.enums.Flag;
-import de.samply.bbmri.negotiator.jooq.tables.Person;
 import de.samply.bbmri.negotiator.rest.Directory;
 import eu.bbmri.eric.csit.service.negotiator.sync.directory.dto.DirectoryBiobank;
 import eu.bbmri.eric.csit.service.negotiator.sync.directory.dto.DirectoryCollection;
@@ -74,258 +69,7 @@ public class DbUtil {
     private final static Logger logger = LoggerFactory.getLogger(DbUtil.class);
     private static DatabaseListMapper databaseListMapper = new DatabaseListMapper();
 
-    /**
-     * Sets the field for starting negotiation for a query to true.
-     * @param config JOOQ configuration
-     * @param queryId id of the query
-     */
-    public static String startNegotiation(Config config, Integer queryId){
-        config.dsl().update(Tables.QUERY)
-                .set(Tables.QUERY.NEGOTIATION_STARTED_TIME, new Timestamp(new Date().getTime()))
-                .where(Tables.QUERY.ID.eq(queryId))
-                .execute();
-        try {
-            config.commit();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-
-    /**
-     * Sets the field for starting negotiation for a query to true.
-     * @param config JOOQ configuration
-     * @param queryId id of the query
-     */
-    public static String restNegotiation(Config config, Integer queryId){
-        config.dsl().execute("UPDATE query SET negotiation_started_time=null WHERE id=" + queryId + ";");
-        try {
-            config.commit();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-
-    /**
-     * Gets the time when first valid query was created in the negotiator.
-     * @param config JOOQ configuration
-     * @return Timestamp timestamp of query
-     */
-    public static Timestamp getFirstQueryCreationTime(Config config){
-        Record1<Timestamp> result = config.dsl()
-                .select(Tables.QUERY.QUERY_CREATION_TIME)
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.VALID_QUERY.eq(true))
-                .orderBy(Tables.QUERY.QUERY_CREATION_TIME.asc())
-                .fetchAny();
-
-        if (result == null){
-            return null;
-        }
-
-        Timestamp timestamp =  result.value1();
-        return timestamp;
-    }
-
-    /**
-     * Gets all the valid queries that entered the negotiator after the given timestamp.
-     * @param config JOOQ configuration
-     * @param timestamp
-     * @return List<QueryDetail> list of queries
-     */
-    public static List<QueryDetail> getAllNewQueries(Config config, Timestamp timestamp) {
-        Result<Record> result = config.dsl()
-                .select(Tables.QUERY.TITLE.as("query_title"))
-                .select(Tables.QUERY.TEXT.as("query_text"))
-                .select(Tables.QUERY.ID.as("query_id"))
-                .select(Tables.QUERY.QUERY_XML.as("query_xml"))
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.VALID_QUERY.eq(true))
-                .and( Tables.QUERY.QUERY_CREATION_TIME.ge(timestamp))
-                .fetch();
-
-        return MappingListDbUtil.getQueryDetails(result);
-    }
-
-    /**
-     * Gets the time when the last connector request was made for the queries.
-     * @param config JOOQ configuration
-     * @param collectionId collection id of the connector
-     * @return  Timestamp of last request
-     */
-    public static Timestamp getLastNewQueryTime(Config config, String collectionId) {
-        Record1<Timestamp> result = config.dsl()
-                .select(Tables.CONNECTOR_LOG.LAST_QUERY_TIME)
-                .from(Tables.CONNECTOR_LOG)
-                .where(Tables.CONNECTOR_LOG.DIRECTORY_COLLECTION_ID.eq(collectionId))
-                .and (Tables.CONNECTOR_LOG.LAST_QUERY_TIME.isNotNull())
-                .orderBy(Tables.CONNECTOR_LOG.LAST_QUERY_TIME.desc())
-                .fetchAny();
-
-        if (result == null){
-            return null;
-        }
-
-        Timestamp timestamp = result.value1();
-        return timestamp;
-    }
-
-    /**
-     * Logs the time when the connector request was made for the queries.
-     *
-     * @param config JOOQ configuration
-     * @param collectionId The collection directoryID
-     * @return 1:ok, 0:error, -1:collectionId unknown
-     */
-    public static int logGetQueryTime(Config config, String collectionId) {
-        try { ConnectorLogRecord connectorLogRecord = config.dsl().newRecord(Tables.CONNECTOR_LOG);
-              connectorLogRecord.setDirectoryCollectionId(collectionId);
-              connectorLogRecord.setLastQueryTime(new Timestamp(new Date().getTime()));
-              connectorLogRecord.store();
-              return 1;
-        } catch (Exception e) {
-            if(e instanceof DataAccessException) {
-                if(e.getMessage().contains("is not present in table \"collection\"")) {
-                    logger.error("Collection with name "+collectionId+" unknown!");
-                    return -1;
-                } else {
-                    logger.error("Data Access Exception: " + e.getMessage());
-                    e.printStackTrace();
-                }
-            } else {
-                e.printStackTrace();
-            }
-        }
-
-        return 0;
-    }
-
-    /**
-     * Get title and text of a query.
-     * @param config JOOQ configuration
-     * @param queryId the query id for which the edit description started
-     * @return QueryRecord object
-     * @throws SQLException
-     */
-    public static QueryRecord getQueryFromId(Config config, Integer queryId) {
-        Record result = config.dsl().select(getFields(Tables.QUERY, "query"))
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.ID.eq(queryId))
-                .fetchOne();
-
-        return MappingDbUtil.mapRecordQueryRecord(result);
-    }
-
-    public static de.samply.bbmri.negotiator.jooq.tables.pojos.Query getQueryFromIdAsQuery(Config config, Integer queryId) {
-        Record result = config.dsl().select(getFields(Tables.QUERY, "query"))
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.ID.eq(queryId))
-                .fetchOne();
-        return MappingDbUtil.mapRequestQuery(result);
-    }
-
     // TODO: Rework Mapping
-    /**
-     * Edits/Updates title, description and jsonText of a query.
-     * @param title title of the query
-     * @param text description of the query
-     * @param queryId the query id for which the editing started
-     * @throws SQLException
-     * @throws IOException
-     * @throws JsonMappingException
-     * @throws JsonParseException
-     */
-    public static void editQuery(Config config, String title, String text, String requestDescription,
-            String jsonText, String ethicsVote, Integer queryId, boolean testRequest) {
-        try {config.dsl().update(Tables.QUERY)
-                .set(Tables.QUERY.TITLE, title)
-                .set(Tables.QUERY.TEXT, text)
-                .set(Tables.QUERY.REQUEST_DESCRIPTION, requestDescription)
-                .set(Tables.QUERY.JSON_TEXT, jsonText)
-                .set(Tables.QUERY.ETHICS_VOTE, ethicsVote)
-                .set(Tables.QUERY.TEST_REQUEST, testRequest)
-                .set(Tables.QUERY.VALID_QUERY, true).where(Tables.QUERY.ID.eq(queryId)).execute();
-
-            /**
-             * Updates the relationship between query and collection.
-             */
-            // TODO: ERROR in Mapper -> Resulting in BiobankID and CollectionID = null
-            QueryDTO queryDTO = Directory.getQueryDTO(jsonText);
-            for(QuerySearchDTO querySearchDTO : queryDTO.getSearchQueries()) {
-                ListOfDirectories listOfDirectories = DbUtilListOfDirectories.getDirectoryByUrl(config, querySearchDTO.getUrl());
-                // collections already saved for this query
-                List<CollectionBiobankDTO> alreadySavedCollectiontsList = getCollectionsForQuery(config, queryId);
-                HashMap<Integer, Boolean> alreadySavedCollections = new HashMap<>();
-                for (CollectionBiobankDTO savedOne : alreadySavedCollectiontsList) {
-                    alreadySavedCollections.put(savedOne.getCollection().getId(), true);
-                }
-
-                if (NegotiatorConfig.get().getNegotiator().getDevelopment().isFakeDirectoryCollections()
-                        && NegotiatorConfig.get().getNegotiator().getDevelopment().getCollectionList() != null) {
-                    logger.info("Faking collections from the directory.");
-                    for (String collection : NegotiatorConfig.get().getNegotiator().getDevelopment().getCollectionList()) {
-                        CollectionRecord dbCollection = getCollection(config, collection, listOfDirectories.getId());
-
-                        if (dbCollection != null) {
-                            if (!alreadySavedCollections.containsKey(dbCollection.getId())) {
-                                addQueryToCollection(config, queryId, dbCollection.getId());
-                                alreadySavedCollections.put(dbCollection.getId(), true);
-                            }
-                        }
-                    }
-                } else {
-                    for (CollectionDTO collection : querySearchDTO.getCollections()) {
-                        CollectionRecord dbCollection = getCollection(config, collection.getCollectionID(), listOfDirectories.getId());
-
-                        if (dbCollection != null) {
-                            if (!alreadySavedCollections.containsKey(dbCollection.getId())) {
-                                addQueryToCollection(config, queryId, dbCollection.getId());
-                                alreadySavedCollections.put(dbCollection.getId(), true);
-                            }
-                        }
-                    }
-                }
-            }
-
-            config.commit();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static void transferQuery(Config config, Integer queryId, Integer researcherId) {
-        de.samply.bbmri.negotiator.jooq.tables.pojos.Person researcher = getPersonDetails(config, researcherId);
-        config.dsl().update(Tables.QUERY)
-                .set(Tables.QUERY.RESEARCHER_ID, researcherId)
-                .set(Tables.QUERY.RESEARCHER_EMAIL, researcher.getAuthEmail())
-                .set(Tables.QUERY.RESEARCHER_NAME, researcher.getAuthName())
-                .set(Tables.QUERY.RESEARCHER_ORGANIZATION, researcher.getOrganization())
-                .where(Tables.QUERY.ID.eq(queryId))
-                .execute();
-    }
-
-    /**
-     * Insert query attachment name
-     * @param queryId
-     * @param attachment
-     * @return the ID of the inserted attachment
-     * @throws SQLException
-     */
-    public static Integer insertQueryAttachmentRecord(Config config, Integer queryId, String attachment, String attachmentType) {
-        Result<QueryAttachmentRecord> result = config.dsl().insertInto(Tables.QUERY_ATTACHMENT)
-                .set(Tables.QUERY_ATTACHMENT.ATTACHMENT, attachment)
-                .set(Tables.QUERY_ATTACHMENT.QUERY_ID, queryId)
-                .set(Tables.QUERY_ATTACHMENT.ATTACHMENT_TYPE, attachmentType)
-                .returning(Tables.QUERY_ATTACHMENT.ID)
-                .fetch();
-
-        if(result == null || result.getValues(Tables.QUERY_ATTACHMENT.ID) == null || result.getValues(Tables
-                .QUERY_ATTACHMENT.ID).size() < 1)
-            return null;
-
-        return result.getValues(Tables.QUERY_ATTACHMENT.ID).get(0);
-    }
 
     public static Integer insertPrivateAttachmentRecord(Config config, Integer queryId, String attachment, String attachmentType, Integer personId, Integer biobank_in_private_chat, Timestamp attachment_time) {
         Result<QueryAttachmentPrivateRecord> result = config.dsl().insertInto(Tables.QUERY_ATTACHMENT_PRIVATE)
@@ -361,37 +105,6 @@ public class DbUtil {
         return result.getValues(Tables.QUERY_ATTACHMENT_COMMENT.ID).get(0);
     }
 
-    public static Integer insertQueryAttachmentRecord(Config config, AttachmentDTO fileDTO) {
-        if(fileDTO.getClass().equals(QueryAttachmentDTO.class)) {
-            QueryAttachmentDTO queryFileDTO = (QueryAttachmentDTO)fileDTO;
-            return insertQueryAttachmentRecord(config, queryFileDTO.getQueryId(), queryFileDTO.getAttachment(), queryFileDTO.getAttachmentType());
-        } else if (fileDTO.getClass().equals(PrivateAttachmentDTO.class)) {
-            PrivateAttachmentDTO privateFileDTO = (PrivateAttachmentDTO)fileDTO;
-            return insertPrivateAttachmentRecord(config, privateFileDTO.getQueryId(), privateFileDTO.getAttachment(), privateFileDTO.getAttachmentType(),
-                    privateFileDTO.getPersonId(), privateFileDTO.getBiobank_in_private_chat(), privateFileDTO.getAttachment_time());
-        } else if (fileDTO.getClass().equals(CommentAttachmentDTO.class)) {
-            CommentAttachmentDTO commentFileDTO = (CommentAttachmentDTO)fileDTO;
-            return insertCommentAttachmentRecord(config, commentFileDTO.getQueryId(), commentFileDTO.getAttachment(),
-                    commentFileDTO.getAttachmentType(), commentFileDTO.getCommentId());
-        } else {
-            logger.error("Error insertQueryAttachmentRecord: No matching Attachment Class.");
-            return null;
-        }
-    }
-
-
-    public static void deleteQueryAttachmentRecord(Config config, Integer queryId, Integer attachment) {
-        config.dsl().delete(Tables.QUERY_ATTACHMENT)
-            .where(Tables.QUERY_ATTACHMENT.QUERY_ID.eq(queryId))
-            .and(Tables.QUERY_ATTACHMENT.ID.eq(attachment))
-            .execute();
-
-        config.dsl().update(Tables.QUERY)
-            .set(Tables.QUERY.NUM_ATTACHMENTS, Tables.QUERY.NUM_ATTACHMENTS.sub(1))
-            .where(Tables.QUERY.ID.eq(queryId))
-            .execute();
-
-    }
 
     public static void deleteCommentAttachment(Config config, Integer commentAttachmentId) {
         config.dsl().delete(Tables.QUERY_ATTACHMENT_COMMENT)
@@ -405,208 +118,6 @@ public class DbUtil {
                 .execute();
     }
 
-    /**
-     * Update number of attachments associated with this query (existing and deleted)
-     * @param numAttachments
-     * @param queryId
-     * @throws SQLException
-     */
-    public static void updateNumQueryAttachments(Config config, Integer queryId, Integer numAttachments) {
-        config.dsl().update(Tables.QUERY)
-                    .set(Tables.QUERY.NUM_ATTACHMENTS, numAttachments)
-                    .where(Tables.QUERY.ID.eq(queryId))
-                    .execute();
-    }
-
-
-    /**
-     * Get the JSON query from the database.
-     * @param config JOOQ configuration
-     * @param queryId the query ID
-     * @return JSON string
-     */
-    public static String getJsonQuery(Config config, Integer queryId) {
-        return config.dsl().selectFrom(Tables.JSON_QUERY)
-                .where(Tables.JSON_QUERY.ID.eq(queryId))
-                .fetchOne(Tables.JSON_QUERY.JSON_TEXT);
-    }
-
-    /**
-     * Get the JSON query from the database.
-     * @param config JOOQ configuration
-     * @param queryId the query ID
-     * @return JSON string
-     */
-    public static String getQuery(Config config, Integer queryId) {
-        return config.dsl().selectFrom(Tables.QUERY)
-                    .where(Tables.QUERY.ID.eq(queryId))
-                    .fetchOne(Tables.QUERY.JSON_TEXT);
-    }
-
-    /**
-     * Get the JSON query from the database.
-     * @param config JOOQ configuration
-     * @param token the negotiator token that also identifies a query. Used for the interaction with the directory
-     * @return JSON string
-     */
-    public static QueryRecord getQuery(Config config, String token) {
-        return config.dsl().selectFrom(Tables.QUERY)
-                .where(Tables.QUERY.NEGOTIATOR_TOKEN.eq(token))
-                .fetchOne();
-    }
-
-    /**
-     * Returns a list of queries with the number of biobanks that commented on that query and the last
-     * time a comment was made
-     * @param config jooq configuration
-     * @param userId the researcher ID
-     * @return
-     */
-    public static List<QueryStatsDTO> getQueryStatsDTOs(Config config, int userId, Set<String> filters) {
-        Person commentAuthor = Tables.PERSON.as("comment_author");
-
-        Condition condition = Tables.QUERY.RESEARCHER_ID.eq(userId);
-
-        if(filters != null && filters.size() > 0) {
-            Condition titleCondition = DSL.trueCondition();
-            Condition textCondition = DSL.trueCondition();
-            Condition nameCondition = DSL.trueCondition();
-
-            for(String filter : filters) {
-                titleCondition = titleCondition.and(Tables.QUERY.TITLE.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-                textCondition = textCondition.and(Tables.QUERY.TEXT.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-                nameCondition = nameCondition.and(commentAuthor.AUTH_NAME.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-
-            }
-            condition = condition.and(titleCondition.or(textCondition).or(nameCondition));
-        }
-
-        Result<Record> records = config.dsl()
-                .select(getFields(Tables.QUERY, "query"))
-                .select(getFields(Tables.PERSON, "person"))
-                .select(Tables.COMMENT.COMMENT_TIME.max().as("last_comment_time"))
-                .select(Tables.COMMENT.ID.countDistinct().as("comment_count"))
-                .select(Tables.PERSON_COMMENT.COMMENT_ID.countDistinct().as("unread_comment_count"))
-                .from(Tables.QUERY)
-                .join(Tables.PERSON, JoinType.LEFT_OUTER_JOIN).on(Tables.PERSON.ID.eq(Tables.QUERY.RESEARCHER_ID))
-                .join(Tables.COMMENT, JoinType.LEFT_OUTER_JOIN).on(Tables.COMMENT.QUERY_ID.eq(Tables.QUERY.ID).and(Tables.COMMENT.STATUS.eq("published")))
-                .join(commentAuthor, JoinType.LEFT_OUTER_JOIN).on(Tables.COMMENT.PERSON_ID.eq(commentAuthor.ID))
-                .join(Tables.PERSON_COMMENT, JoinType.LEFT_OUTER_JOIN)
-                    .on(Tables.PERSON_COMMENT.COMMENT_ID.eq(Tables.COMMENT.ID)
-                            .and(Tables.PERSON_COMMENT.PERSON_ID.eq(Tables.QUERY.RESEARCHER_ID))
-                            .and(Tables.PERSON_COMMENT.READ.eq(false)))
-                .leftOuterJoin(Tables.REQUEST_STATUS)
-                    .on(Tables.REQUEST_STATUS.QUERY_ID.eq(Tables.QUERY.ID)
-                            .and(Tables.REQUEST_STATUS.STATUS.eq("abandoned")))
-                .where(condition).and(Tables.REQUEST_STATUS.STATUS.isNull())
-                .groupBy(Tables.QUERY.ID, Tables.PERSON.ID)
-                .orderBy(Tables.QUERY.QUERY_CREATION_TIME.desc()).fetch();
-
-        return MappingListDbUtil.mapRecordResultQueryStatsDTOList(records);
-    }
-
-    /**
-     * Returns a list of queries for a particular owner, filtered by a search term if such is provided
-     * @param config jooq configuration
-     * @param userId the user ID of the biobank owner
-     * @param filters search term for title and text
-     * @return
-     */
-    public static List<OwnerQueryStatsDTO> getOwnerQueries(Config config, int userId, Set<String> filters, Flag flag, Boolean isTestRequest) {
-    	Person queryAuthor = Tables.PERSON.as("query_author");
-
-    	Condition condition = Tables.PERSON_COLLECTION.PERSON_ID.eq(userId);
-
-    	if(filters != null && filters.size() > 0) {
-            Condition titleCondition = DSL.trueCondition();
-            Condition textCondition = DSL.trueCondition();
-            Condition nameCondition = DSL.trueCondition();
-
-            for(String filter : filters) {
-                titleCondition = titleCondition.and(Tables.QUERY.TITLE.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-    			textCondition = textCondition.and(Tables.QUERY.TEXT.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-   			    nameCondition = nameCondition.and(queryAuthor.AUTH_NAME.likeIgnoreCase("%" + filter.replace("%", "!%") + "%", '!'));
-            }
-    		condition = condition.and(titleCondition.or(textCondition).or(nameCondition));
-    	}
-
-        if (flag != null && flag != Flag.UNFLAGGED) {
-            condition = condition.and(Tables.FLAGGED_QUERY.FLAG.eq(flag));
-        } else {
-            /**
-             * Ignored queries are never selected unless the user is in the ignored folder
-             */
-    		condition = condition.and(Tables.FLAGGED_QUERY.FLAG.ne(Flag.IGNORED).or(Tables.FLAGGED_QUERY.FLAG.isNull()));
-    	}
-
-        condition = condition.and(Tables.QUERY.NEGOTIATION_STARTED_TIME.isNotNull());
-
-        Table<RequestStatusRecord> requestStatusTableStart = Tables.REQUEST_STATUS.as("request_status_table_start");
-        Table<RequestStatusRecord> requestStatusTableAbandon = Tables.REQUEST_STATUS.as("reque_ststatus_table_abandon");
-
-    	Result<Record> fetch = config.dsl()
-				.select(getFields(Tables.QUERY, "query"))
-				.select(getFields(queryAuthor, "person"))
-    			.select(Tables.COMMENT.COMMENT_TIME.max().as("last_comment_time"))
-    			.select(Tables.COMMENT.ID.countDistinct().as("comment_count"))
-                .select(Tables.PERSON_COMMENT.COMMENT_ID.countDistinct().as("unread_comment_count"))
-                .select(DSL.decode().when(Tables.FLAGGED_QUERY.FLAG.isNull(), Flag.UNFLAGGED)
-                        .otherwise(Tables.FLAGGED_QUERY.FLAG).as("flag"))
-    			.from(Tables.QUERY)
-
-    			.join(Tables.QUERY_COLLECTION, JoinType.JOIN)
-    			.on(Tables.QUERY.ID.eq(Tables.QUERY_COLLECTION.QUERY_ID))
-
-                .join(Tables.COLLECTION, JoinType.JOIN)
-                .on(Tables.COLLECTION.ID.eq(Tables.QUERY_COLLECTION.COLLECTION_ID))
-
-                .join(Tables.PERSON_COLLECTION, JoinType.JOIN)
-                .on(Tables.PERSON_COLLECTION.COLLECTION_ID.eq(Tables.COLLECTION.ID))
-
-    			.join(queryAuthor, JoinType.LEFT_OUTER_JOIN)
-    			.on(Tables.QUERY.RESEARCHER_ID.eq(queryAuthor.ID))
-
-    			.join(Tables.COMMENT, JoinType.LEFT_OUTER_JOIN)
-    			.on(Tables.QUERY.ID.eq(Tables.COMMENT.QUERY_ID).and(Tables.COMMENT.STATUS.eq("published")))
-
-    			.join(Tables.FLAGGED_QUERY, JoinType.LEFT_OUTER_JOIN)
-    			.on(Tables.QUERY.ID.eq(Tables.FLAGGED_QUERY.QUERY_ID).and(Tables.FLAGGED_QUERY.PERSON_ID.eq(Tables.PERSON_COLLECTION.PERSON_ID)))
-
-                .join(requestStatusTableStart, JoinType.JOIN)
-                .on(Tables.QUERY.ID.eq(requestStatusTableStart.field(Tables.REQUEST_STATUS.QUERY_ID))
-                        .and(requestStatusTableStart.field(Tables.REQUEST_STATUS.STATUS).eq("started")))
-
-                .join(requestStatusTableAbandon, JoinType.LEFT_OUTER_JOIN)
-                .on(Tables.QUERY.ID.eq(requestStatusTableAbandon.field(Tables.REQUEST_STATUS.QUERY_ID))
-                        .and(requestStatusTableAbandon.field(Tables.REQUEST_STATUS.STATUS).eq("abandoned")))
-
-                .join(Tables.PERSON_COMMENT, JoinType.LEFT_OUTER_JOIN)
-                .on(Tables.PERSON_COMMENT.COMMENT_ID.eq(Tables.COMMENT.ID)
-                        .and(Tables.PERSON_COMMENT.PERSON_ID.eq(Tables.PERSON_COLLECTION.PERSON_ID))
-                        .and(Tables.PERSON_COMMENT.READ.eq(false)))
-
-    			.where(condition).and(requestStatusTableAbandon.field(Tables.REQUEST_STATUS.STATUS).isNull())
-                .and(Tables.QUERY.TEST_REQUEST.eq(isTestRequest))
-    			.groupBy(Tables.QUERY.ID, queryAuthor.ID, Tables.FLAGGED_QUERY.PERSON_ID, Tables.FLAGGED_QUERY.QUERY_ID)
-    			.orderBy(Tables.QUERY.QUERY_CREATION_TIME.desc()).fetch();
-
-        return MappingListDbUtil.mapRecordResultOwnerQueryStatsDTOList(fetch);
-    }
-
-    /**
-     * Returns a list of QueryAttachmentDTO for a specific query.
-     * @param config
-     * @param queryId
-     * @return List<QueryAttachmentDTO>
-     */
-    public static List<QueryAttachmentDTO> getQueryAttachmentRecords(Config config, int queryId) {
-        Result<Record> result = config.dsl()
-                .select(getFields(Tables.QUERY_ATTACHMENT, "query_attachment"))
-                .from(Tables.QUERY_ATTACHMENT)
-                .where(Tables.QUERY_ATTACHMENT.QUERY_ID.eq(queryId))
-                .orderBy(Tables.QUERY_ATTACHMENT.ID.asc()).fetch();
-        return MappingListDbUtil.mapRecordsQueryAttachmentDTO(result);
-    }
 
     public static List<PrivateAttachmentDTO> getPrivateAttachmentRecords(Config config, int queryId) {
         Result<Record> result = config.dsl()
@@ -948,7 +459,7 @@ public class DbUtil {
      * @param directoryId directory collection ID
      * @return
      */
-    private static CollectionRecord getCollection(Config config, String directoryId, int listOfDirectoryId) {
+    public static CollectionRecord getCollection(Config config, String directoryId, int listOfDirectoryId) {
         return config.dsl().selectFrom(Tables.COLLECTION)
                 .where(Tables.COLLECTION.DIRECTORY_ID.eq(directoryId))
                 .and(Tables.COLLECTION.LIST_OF_DIRECTORIES_ID.eq(listOfDirectoryId))
@@ -1229,7 +740,7 @@ public class DbUtil {
      * @param queryId the query id which will be associated with a collection
      * @param collectionId the collection id which will be associated with the query
      */
-    private static void addQueryToCollection(Config config, Integer queryId, Integer collectionId) {
+    public static void addQueryToCollection(Config config, Integer queryId, Integer collectionId) {
         addQueryToCollection(config, queryId, collectionId, false);
     }
