+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/mapping/mapper/OwnerQueryStatsDTOMapper.java
@@ -1,0 +1,8 @@
+package eu.bbmri.eric.csit.service.negotiator.mapping.mapper;
+
+public class OwnerQueryStatsDTOMapper {
+
+    private OwnerQueryStatsDTOMapper() {}
+
+    public map()
+}
