+++ src/test/java/eu/bbmri/eric/csit/service/negotiator/mapping/DatabaseObjectMapperTest.java
@@ -2,7 +2,7 @@ package eu.bbmri.eric.csit.service.negotiator.mapping;
 
 import de.samply.bbmri.negotiator.jooq.tables.pojos.ListOfDirectories;
 import de.samply.bbmri.negotiator.jooq.tables.pojos.Person;
-import de.samply.bbmri.negotiator.jooq.tables.pojos.Query;;
+import de.samply.bbmri.negotiator.jooq.tables.pojos.Query;
 import org.jooq.Record;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -12,7 +12,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.sql.Timestamp;
